Add multiple utility classes and supporting schemas
A set of utility and schema classes are added to the project. This includes classes for generating hashed values, mock HTTP responses, and initializing user data, among others. Also, schemas for describing objects such as User and SObject have been introduced. This commit also includes changes to project settings and the addition of a .gitignore file.