Add SelectionSort and BubbleSort, modify BinarySort and LinearSort
This commit focuses on the addition of SelectionSort and BubbleSort implementations as well as modifications to existing BinarySort and LinearSort. It also includes the deletion of ShellSort and refactoring changes in .idea and other sorting files. While making additions and updates, the emphasis has been on improving efficiency and code clarity.